+++ src/repo_analyzer/render.clj
@@ -175,6 +175,29 @@
       )))
 
 
+(defn create-file-change-statistics-html
+  [file-change-statistics]
+  (html
+    [:h2 "File change statistics"]
+    [:table {:class "table table-striped"}
+     [:thead
+      [:tr
+       [:th {:scope "col"} "Filename"]
+       [:th {:scope "col"} "Nr of edits"]
+       ]
+      ]
+     [:tbody
+      (map #(vector
+              :tr
+              [:td (first %)]
+              [:td (count (:edits (second %)))]
+              ) file-change-statistics)
+      ]
+     ]
+    )
+  )
+
+
 (defn create-commit-statistics
   "Creates commit statistics HTML and subpages. Returns the created HTML"
   [analysis base-path]
@@ -191,6 +214,7 @@
         line-chart-labels (string/join "," (map #(string/join ["\"" (first %) "\""]) (get-in analysis [:commit-statistics :time-distribution])))
         line-chart-data-authored (string/join "," (map #(:authored (second %)) (get-in analysis [:commit-statistics :time-distribution])))
         line-chart-data-committed (string/join "," (map #(:committed (second %)) (get-in analysis [:commit-statistics :time-distribution])))
+        file-change-statistics-html (create-file-change-statistics-html (:file-change-statistics analysis))
         ]
     (create-site all-commits-filename "All commits" commit-list-html)
     (create-site self-commits-filename "Self committed commits" self-commit-list-html)
@@ -223,6 +247,7 @@
        (:count (:commit-statistics analysis)) "(" (get-in analysis [:commit-statistics :committed-by-different-dev :percentage]) "%)"
        [:a {:href "different-committer-list.html"} " See list of all commits where author and committer are different"]
        ]
+      file-change-statistics-html
       )
     ))